+++ src/controller.ts
@@ -1,12 +1,12 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import * as fs from 'fs';
-import { SDKClient } from './util';
-import {yuqueClone} from './yuque';
-import { YuqueOutlineProvider } from './yuqueOutline';
-import { DocumentNode } from './documentNode';
 import * as open from 'open';
 import * as open_darwin from 'mac-open';
+import { SDKClient } from './util';
+import {updateOrCreateTOC, yuqueClone} from './yuque';
+import { YuqueOutlineProvider } from './yuqueOutline';
+import { DocumentNode } from './documentNode';
 
 // decide what os should be used
 // possible node values 'darwin', 'freebsd', 'linux', 'sunos' or 'win32'
@@ -26,12 +26,28 @@ export class YuqueController {
                 this._yuqueOutlineProvider.refresh();
             }
         );
+        vscode.commands.registerCommand('yuqueCli.updateTOC',
+            async () => {
+                await this.updateTOC();
+                this._yuqueOutlineProvider.refresh();
+            }
+        );
+
         vscode.commands.registerCommand('yuqueCli.fetchDocument',
             () => this.fetchDocument());
+        vscode.commands.registerCommand('yuqueCli.openDocument',
+            () => this.openDocument());
         vscode.commands.registerCommand('yuqueCli.updateDocument',
             () => this.updateDocument());
         vscode.commands.registerCommand('yuqueCli.createDocument',
             () => this.createDocument());
+        vscode.commands.registerCommand('yuqueCli.deleteDocument',
+            async () => {
+                await this.deleteDocument();
+                await this.updateTOC();
+                this._yuqueOutlineProvider.refresh();
+            }
+        );
         vscode.commands.registerCommand('yuqueCli.onDocumentClicked',
             (node: DocumentNode) => this.onYuqueDocumentClicked(node));
     }
@@ -46,13 +62,17 @@ export class YuqueController {
             let document: any = await SDKClient.docs.get(
                 {namespace: namespace, slug: this._lastIDClicked.id, data: {raw: 1}});
             let documentBody = document.body_draft || document.body;
-            let folders = vscode.workspace.workspaceFolders;
-            if (folders.length === 1) {
-                let docPath = path.join(folders[0].uri.fsPath, this._lastIDClicked.id.toString() + '.md');
+
+            let workspacePath = this.getWorkspaceFolder();
+            if (workspacePath) {
+                let docPath = path.join(workspacePath, this._lastIDClicked.id.toString() + '.md');
                 fs.writeFile(docPath, documentBody, {}, function(err) {
-                    console.log(err);
+                    vscode.window.showErrorMessage(err.message);
                 });
             }
+            vscode.window.showInformationMessage(`Fetch ${this._lastIDClicked.label} successfully`);
+        } else {
+            vscode.window.showWarningMessage('Please left click TreeItem first');
         }
     }
 
@@ -79,14 +99,16 @@ export class YuqueController {
                     open(url);
                 }
             }
+        } else {
+            vscode.window.showErrorMessage('Title must not be null');
         }
     }
 
     async updateDocument() {
         if (this._lastIDClicked) {
-            let folders = vscode.workspace.workspaceFolders;
-            if (folders.length === 1) {
-                let docPath = path.join(folders[0].uri.fsPath, this._lastIDClicked.id.toString() + '.md');
+            let workspacePath = this.getWorkspaceFolder();
+            if (workspacePath) {
+                let docPath = path.join(workspacePath, this._lastIDClicked.id.toString() + '.md');
                 const documentBody = fs.readFileSync(docPath, 'utf-8');
                 SDKClient.docs.update({
                     namespace: this._yuqueOutlineProvider.namespace(),
@@ -95,7 +117,53 @@ export class YuqueController {
                         body: documentBody
                     }
                 });
+                vscode.window.showInformationMessage(`Update ${this._lastIDClicked.label} successfully`);
             } 
+        } else {
+            vscode.window.showWarningMessage('Please left click TreeItem first');
         }
+    }
+
+    async openDocument() {
+        if (this._lastIDClicked) {
+            let workspacePath = this.getWorkspaceFolder();
+            if (workspacePath) {
+                let docPath = path.join(workspacePath, this._lastIDClicked.id.toString() + '.md');
+                if (fs.existsSync(docPath)) {
+                    let uri = vscode.Uri.file(docPath);
+                    vscode.workspace.openTextDocument(uri).then(document => vscode.window.showTextDocument(document));
+                } else {
+                    vscode.window.showErrorMessage('File not fetched');
+                }
+            }
+        } else {
+            vscode.window.showWarningMessage('Please left click TreeItem first');
+        }
+    }
+
+    async deleteDocument() {
+        if (this._lastIDClicked) {
+            await SDKClient.docs.delete({
+                namespace: this._yuqueOutlineProvider.namespace(),
+                id: this._lastIDClicked.id
+            });
+            vscode.window.showInformationMessage(`Delete ${this._lastIDClicked.label} successfully`);
+        } else {
+            vscode.window.showWarningMessage('Please left click TreeItem first');
+        }
+    }
+
+    async updateTOC() {
+        await updateOrCreateTOC(this._yuqueOutlineProvider.namespace());
+    }
+
+    private getWorkspaceFolder(): string | undefined {
+        let folders = vscode.workspace.workspaceFolders;
+        if (folders.length === 1) {
+            return folders[0].uri.fsPath;
+        } else {
+            vscode.window.showErrorMessage('YuqueCli is not supported for multiworkspaces');
+        }
+
     }
 }